\name{diag-dotboxplot-package}
\alias{diag-dotboxplot-package}
\alias{diag-dotboxplot}
\docType{package}
\title{
\packageTitle{diag-dotboxplot}
}
\description{
\packageDescription{diag-dotboxplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diag-dotboxplot}
\packageIndices{diag-dotboxplot}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{diag-dotboxplot}

Maintainer: \packageMaintainer{diag-dotboxplot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
